requirements_IONQ.txt
qiskit==1.4.2
qiskit-algorithms==0.4.0
qiskit-optimization==0.7.0
qiskit-ionq==0.5.0
python-dotenv